Make the mapreduce openmp examples more uniform

+++ examples/mapreduce/cpp/openacc/optimized.cpp
@@ -7,7 +7,7 @@ using namespace std;
 int main() {
   int num = 100000000;
   int *elements = new int[num];
-  long long res = 0;
+  long res = 0;
 
   timer time;
 
@@ -18,6 +18,7 @@ int main() {
     for (int i = 0; i < num; i++) {
       elements[i] = i;
     }
+
     #pragma acc parallel loop reduction(+ : res)
     for (int i = 0; i < num; i++) {
       res += elements[i];
